+++ CHANGELOG.md
@@ -6,7 +6,9 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [x.x.x]
 ### Added
+ - Reinstated tooltips to display additional information on genetic elements
 ### Changed
+ - Use popper for positioning tooltips
 ### Fixed
 
 ## [1.2.0]